fixed cr comments
