+++ packages/adapter-components/src/config/shared.ts
@@ -19,7 +19,7 @@ import {
 } from '@salto-io/adapter-api'
 import { createMatchingObjectType } from '@salto-io/adapter-utils'
 import { RequestConfig, RequestDefaultConfig } from './request'
-import { TransformationConfig, TransformationDefaultConfig } from './transformation'
+import type { TransformationConfig, TransformationDefaultConfig } from './transformation'
 
 export type TypeConfig<T extends TransformationConfig = TransformationConfig> = {
   request?: RequestConfig
@@ -128,14 +128,3 @@ export const getConfigWithDefault = <
       defaultConfig,
     )
   )
-
-export const getTypeTransformationConfig = (
-  typeName: string,
-  typeConfig: Record<string, TypeConfig>,
-  typeDefaultConfig: TypeDefaultsConfig
-): TransformationConfig => (
-  getConfigWithDefault(
-    typeConfig[typeName]?.transformation,
-    typeDefaultConfig.transformation,
-  )
-)
