+++ packages/adapter-components/src/config/transformation.ts
@@ -18,6 +18,7 @@ import { ElemID, ObjectType, BuiltinTypes, CORE_ANNOTATIONS,
   FieldDefinition, ListType, RestrictionAnnotationType } from '@salto-io/adapter-api'
 import { types } from '@salto-io/lowerdash'
 import { findDuplicates } from './validation_utils'
+import { getConfigWithDefault, TypeConfig, TypeDefaultsConfig } from './shared'
 
 export const DATA_FIELD_ENTIRE_OBJECT = '.'
 
@@ -29,7 +30,6 @@ type FieldToAdjustType = {
   fieldName: string
   fieldType?: string
 }
-
 export type FieldToOmitType = FieldToAdjustType
 export type FieldToHideType = FieldToAdjustType
 export type FieldTypeOverrideType = {
@@ -202,3 +202,14 @@ export const validateTransoformationConfig = (
     _.mapValues(configMap, c => c.standaloneFields),
   )
 }
+
+export const getTypeTransformationConfig = (
+  typeName: string,
+  typeConfig: Record<string, TypeConfig>,
+  typeDefaultConfig: TypeDefaultsConfig
+): TransformationConfig => (
+  getConfigWithDefault(
+    typeConfig[typeName]?.transformation,
+    typeDefaultConfig.transformation,
+  )
+)
