+++ packages/adapter-components/src/elements/type_elements.ts
@@ -15,12 +15,14 @@
 * limitations under the License.
 */
 import _ from 'lodash'
-import { FieldDefinition, Field, CORE_ANNOTATIONS, TypeElement, isObjectType, isContainerType, getDeepInnerType } from '@salto-io/adapter-api'
+import { FieldDefinition, Field, CORE_ANNOTATIONS, TypeElement, isObjectType, isContainerType,
+  getDeepInnerType, ObjectType, BuiltinTypes, createRestriction, createRefToElmWithValue, PrimitiveType, LIST_ID_PREFIX, GENERIC_ID_PREFIX, GENERIC_ID_SUFFIX, MAP_ID_PREFIX, ListType, MapType, isEqualElements } from '@salto-io/adapter-api'
 import { logger } from '@salto-io/logging'
 import { values, collections } from '@salto-io/lowerdash'
-import { FieldToHideType } from '../config/transformation'
+import { FieldToHideType, FieldTypeOverrideType, getTypeTransformationConfig } from '../config/transformation'
 import { SUBTYPES_PATH, TYPES_PATH } from './constants'
 import { getSubtypes } from './subtypes'
+import { TypeConfig, TypeDefaultsConfig } from '../config/shared'
 
 const { awu } = collections.asynciterable
 
@@ -82,4 +84,99 @@ export const filterTypes = async (
     .forEach(t => { t.path = [adapterName, TYPES_PATH, SUBTYPES_PATH, t.elemID.name] })
 
   return [...relevantTypes, ...subtypes]
+}
+
+/**
+ * Helper function for creating the right type element from the type name specified in the config.
+ * This can be called recursively to generate the right type from existing types, potentially
+ * wrapped in containers - such as List<Map<List<someTypeName>>>
+ */
+const getContainerForType = (typeName: string): {
+  container: 'list' | 'map'
+  typeNameSubstring: string
+} | undefined => {
+  if (
+    typeName.toLowerCase().startsWith(`${LIST_ID_PREFIX.toLowerCase()}${GENERIC_ID_PREFIX}`)
+    && typeName.endsWith(GENERIC_ID_SUFFIX)
+  ) {
+    return {
+      container: 'list',
+      typeNameSubstring: typeName.substring(
+        LIST_ID_PREFIX.length + GENERIC_ID_PREFIX.length,
+        typeName.length - GENERIC_ID_SUFFIX.length,
+      ),
+    }
+  }
+  if (
+    typeName.toLowerCase().startsWith(MAP_ID_PREFIX.toLowerCase())
+    && typeName.endsWith(GENERIC_ID_SUFFIX)
+  ) {
+    return {
+      container: 'map',
+      typeNameSubstring: typeName.substring(
+        MAP_ID_PREFIX.length + GENERIC_ID_PREFIX.length,
+        typeName.length - GENERIC_ID_SUFFIX.length,
+      ),
+    }
+  }
+  return undefined
+}
+
+export const fixFieldTypes = (
+  definedTypes: Record<string, ObjectType>,
+  typeConfig: Record<string, TypeConfig>,
+  typeDefaultConfig: TypeDefaultsConfig,
+  toPrimitiveType: (val: string) => PrimitiveType
+): void => {
+  const toTypeWithContainers = (typeName: string): TypeElement => {
+    const containerDetails = getContainerForType(typeName)
+    if (containerDetails?.container === 'list') {
+      return new ListType(toTypeWithContainers(containerDetails.typeNameSubstring))
+    }
+    if (containerDetails?.container === 'map') {
+      return new MapType(toTypeWithContainers(containerDetails.typeNameSubstring))
+    }
+    const type = definedTypes[typeName] ?? toPrimitiveType(typeName)
+    if (isEqualElements(type, BuiltinTypes.UNKNOWN) && typeName.toLowerCase() !== 'unknown') {
+      log.error('could not find type %s, falling back to unknown', typeName)
+    }
+    return type
+  }
+
+  Object.keys(typeConfig)
+    .filter(typeName =>
+      getTypeTransformationConfig(
+        typeName, typeConfig, typeDefaultConfig
+      ).fieldTypeOverrides !== undefined)
+    .forEach(typeName => {
+      const type = definedTypes[typeName]
+      if (type === undefined) {
+        log.warn('type %s not found, cannot override its field types', typeName)
+        return
+      }
+      const fieldTypeOverrides = getTypeTransformationConfig(
+        typeName, typeConfig, typeDefaultConfig,
+      ).fieldTypeOverrides as FieldTypeOverrideType[]
+      fieldTypeOverrides.forEach(({ fieldName, fieldType, restrictions }) => {
+        const field = type.fields[fieldName]
+        const newFieldType = toTypeWithContainers(fieldType)
+        const annotations = restrictions ? {
+          [CORE_ANNOTATIONS.RESTRICTION]: createRestriction(restrictions),
+        } : {}
+        if (field === undefined) {
+          log.debug(
+            'Creating field type for %s.%s with type %s with restrictions %o',
+            typeName, fieldName, newFieldType.elemID.name, restrictions
+          )
+          type.fields[fieldName] = new Field(type, fieldName, newFieldType, annotations)
+        } else {
+          log.debug(
+            'Modifying field type for %s.%s from %s to %s with restrictions %o',
+            typeName, fieldName, field.refType.elemID.name, newFieldType.elemID.name, restrictions
+          )
+          field.refType = createRefToElmWithValue(newFieldType)
+          field.annotations = { ...field.annotations, ...annotations }
+        }
+      })
+    })
 }
